chore: refactor into modules
