+++ src/modules/layers/Folders/NodeIcon.tsx
