+++ src/modules/text/types.ts
@@ -1,0 +1,6 @@
+export type Font = {
+	family: string;
+	files: {
+		[key: string]: string;
+	};
+};
