+++ src/styles/menu.ts
@@ -1,0 +1,7 @@
+import { createStyles } from '@mantine/core';
+
+export const useMenuStyles = createStyles(() => ({
+	item: {
+		fontSize: 12,
+	},
+}));
